Extract ICSCalendarLink to Indico components
