+++ indico/web/client/js/react/components/ICSCalendarLink.jsx
